// src/types/property.type.ts
export interface Region {
  country: string;
  postCode: string;
  state: string;
  suburb: string;
  streetName: string;
  streetNumber: string;
  unit: string;
  buildingName: string;
  longitude: number;
  latitude: number;
}

// 공통 옵션 아이템
export interface OptionItem {
  optionItemId: number;
  itemName: string;
  isActive: boolean;
  isCustom: boolean;
  categoryId: number;
  categoryName: string;
  parentItemId: number | null;
}

// 공통 비용 정보
export interface CostDetails {
  weeklyCost: number;
  costDescription: string;
  deposit: number;
  keyDeposit: number;
  depositAdjustable: boolean;
  billIncluded: boolean;
}

// 공통 거주 조건
export interface LivingConditions {
  noticePeriodWeeks: number;
  minimumStayWeeks: number;
  contractTerms: string;
  contractExtendable: boolean;
}

// 공통 입주 가능 정보
export interface MoveInInfo {
  availableFrom: string;
  availableTo: string;
  immediate: boolean;
  negotiable: boolean;
}

// 쉐어 매물 전용 내부 정보
interface InternalDetailsForShare {
  internalArea: number;
  totalArea: number;
  totalResidents: number;
  totalBathUser: number;
  totalFloors: number;
  propertyFloor: number;
}

// 렌트 매물 전용 내부 정보
interface InternalDetailsForRent {
  internalArea: number;
  totalArea: number;
  numberOfRoom: number;
  numberOfBath: number;
  totalFloors: number;
  propertyFloors: number;
}

interface HostSummary {
  id: number;
  profileImage: string;
  nickname: string;
  verified: boolean;
}

interface MetaInfo {
  owner: boolean;
  wished: boolean;
}

// 공통 매물 베이스 타입
interface BaseProperty {
  id: number;
  memberId: number;
  createdAt: string;
  lastModifiedAt: string;
  optionItems: OptionItem[];
  displayStatus: string;
  tradeStatus: string;
  genderPreference: string;
  lgbtAvailable: boolean;
  region: Region;
  photoUrls: string[];
  thumbnailUrl: string | null;
  costDetails: CostDetails;
  livingConditions: LivingConditions;
  moveInInfo: MoveInInfo;
  parkingOption: string;
  description: string;
  wishCount: number;
  metaInfo?: MetaInfo;
  hostSummary?: HostSummary;
}

// 쉐어 매물 타입
export interface ShareProperty extends BaseProperty {
  kind: "SHARE";
  sharePropertySubType: string;
  internalDetails: InternalDetailsForShare;
  capacityShare: string;
}

// 렌트 매물 타입
export interface RentProperty extends BaseProperty {
  kind: "RENT";
  rentPropertySubType: string;
  internalDetails: InternalDetailsForRent;
  capacityRent: string;
  isRealEstateIntervention: string;
}